add endpoint for update profile attribute

+++ handler/handler_profile.go
@@ -24,7 +24,7 @@ func (h *Handler) createProfile(c *gin.Context) {
 		return
 	}
 
-	if firebaseUid != user.FirebaseUid {
+	if user.FirebaseUid != firebaseUid {
 		c.JSON(http.StatusBadRequest, failure(e.ErrInvalidParams))
 		return
 	}
@@ -72,6 +72,29 @@ func (h *Handler) getProfile(c *gin.Context) {
 	c.JSON(http.StatusOK, user)
 }
 
+func (h *Handler) isAvailable(c *gin.Context) {
+	request := &models.Request{}
+	err := c.ShouldBindJSON(&request)
+	if err != nil {
+		c.JSON(http.StatusBadRequest, failure(err))
+		return
+	}
+
+	if request.Username == "" {
+		c.JSON(http.StatusBadRequest, failure(e.ErrNoParams))
+		return
+	}
+
+	uid, err := h.index.GetUidByUsername(request.Username)
+	if uid != "" {
+		c.JSON(http.StatusOK, gin.H{"available": false})
+	} else if errors.Is(err, e.ErrNotFound) {
+		c.JSON(http.StatusOK, gin.H{"available": true})
+	} else {
+		c.JSON(http.StatusInternalServerError, failure(err))
+	}
+}
+
 func (h *Handler) updateProfile(c *gin.Context) {
 	firebaseUid := c.GetString("firebase_uid")
 	if firebaseUid == "" {
@@ -86,7 +109,7 @@ func (h *Handler) updateProfile(c *gin.Context) {
 		return
 	}
 
-	if firebaseUid != user.FirebaseUid {
+	if user.FirebaseUid != firebaseUid {
 		c.JSON(http.StatusBadRequest, failure(e.ErrInvalidParams))
 		return
 	}
@@ -104,7 +127,13 @@ func (h *Handler) updateProfile(c *gin.Context) {
 	}
 }
 
-func (h *Handler) deleteProfileAttribute(c *gin.Context) {
+func (h *Handler) updateProfileAttribute(c *gin.Context) {
+	firebaseUid := c.GetString("firebase_uid")
+	if firebaseUid == "" {
+		c.JSON(http.StatusBadRequest, failure(e.ErrNoParams))
+		return
+	}
+
 	request := &models.Request{}
 	err := c.ShouldBindJSON(&request)
 	if err != nil {
@@ -118,6 +147,55 @@ func (h *Handler) deleteProfileAttribute(c *gin.Context) {
 	}
 
 	ctx := c.Request.Context()
+	profile, err := h.db.GetProfile(ctx, firebaseUid)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, failure(err))
+		return
+	}
+
+	if profile.Uid != request.Uid {
+		c.JSON(http.StatusBadRequest, failure(e.ErrInvalidParams))
+		return
+	}
+
+	err = h.db.UpdateProfileAttribute(ctx, request.Uid, request.Attribute, request.Value)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, failure(err))
+		return
+	}
+}
+
+func (h *Handler) deleteProfileAttribute(c *gin.Context) {
+	firebaseUid := c.GetString("firebase_uid")
+	if firebaseUid == "" {
+		c.JSON(http.StatusBadRequest, failure(e.ErrNoParams))
+		return
+	}
+
+	request := &models.Request{}
+	err := c.ShouldBindJSON(&request)
+	if err != nil {
+		c.JSON(http.StatusBadRequest, failure(err))
+		return
+	}
+
+	if request.Uid == "" || request.Attribute == "" || request.Value == "" {
+		c.JSON(http.StatusBadRequest, failure(e.ErrNoParams))
+		return
+	}
+
+	ctx := c.Request.Context()
+	profile, err := h.db.GetProfile(ctx, firebaseUid)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, failure(err))
+		return
+	}
+
+	if profile.Uid != request.Uid {
+		c.JSON(http.StatusBadRequest, failure(e.ErrInvalidParams))
+		return
+	}
+
 	err = h.db.DeleteProfileAttribute(ctx, request.Uid, request.Attribute, request.Value)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, failure(err))
